CinemaMashine && Character controller
Made the movement of the player to flow the trajectory of the camera. However for this to be possible I had to change the player movement form Rogibody to Character Controller. Also I used a "Free Lock Camera " of cine machine. All of this was tested inside a new SandBox (scene) . Inside next commit I'll try and do the same but with the rigbody we had since the beginning and also the other cineMashine camera we added.